backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-RESTful==0.3.10
mysql-connector-python==8.2.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Marshmallow==1.1.0
marshmallow-sqlalchemy==0.29.0
flask-jwt-extended==4.6.0
email_validator==2.1.0
qrcode==7.3.1
Pillow==10.2.0
python-dotenv==1.0.0
